Question
A prime number is a number that is divisible by only two numbers, 1 and itself. 1 is neither a prime number nor a composite number. Hence, 2 is the first prime number, 3 is the second prime number and so on..

Your task is to write a program that takes as input an integer N and prints the Nth prime number.

Input
Input consists of a single integer N

Output
Print the Nth prime number

Sample Input 0

1
Sample Output 0

2
Sample Input 1

2
Sample Output 1

3